+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 
 ### Changed
+- Values for "Height", "Angle", "Pitch", and coordinates are now input fields. Users can adjust values using arrow keys.
+- The range for height has been increased to 700'000m.
 
 - Config is loaded from the frontend at runtime now.